refact: organize UI namespace and directories
