+++ Assets/Scripts/UI/Components/City.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc8a3a3a1247cb044b3bcd8e7df8222f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
