+++ Assets/Scripts/UI/Components/City/CityLabelManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3fdd052171259f044b54aba60f2be85e
+guid: f701d6f8b4d40e341a03d82e065bbf3d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
