+++ Assets/Scripts/UI/Deprecated.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fa39d4b5b39fa8547ba5435ec3cacf67
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
