+++ Assets/Scripts/UI/Deprecated/Events.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cce0485a73c73734b867531616346b92
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
