+++ Assets/Scripts/UI/Deprecated/SettlerPanelManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ba8e29015e606c1478e8643ab85f2a15
+guid: 6f1c4650325f0b5418aa6c13d57422ac
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
