+++ Assets/Scripts/UI/Deprecated/UIManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 14907e9ef84a1204e8ddf6a73e9218d3
+guid: 0c2c66f11951ac8499f356c8173bace8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
