+++ Assets/Scripts/UI/Events.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0370b1fdabb4c984da4476d595beb84b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
