+++ Assets/Scripts/UI/Events/CreateCityEvent.cs
@@ -1,0 +1,26 @@
+// using System;
+// using UnityEngine;
+// using Unity.Entities;
+// using Unity.Mathematics;
+// using Unity.Rendering;
+// using Unity.Transforms;
+// using Unity.Collections;
+
+// namespace GameUI
+// {
+//     public class CreateCityEvent : MonoBehaviour {
+//         SettlerPanelManager settlerPanelManager;
+//         private void Start() {
+//             settlerPanelManager = GetComponent<SettlerPanelManager>();
+//             settlerPanelManager.OnCreateCity += OnCreateCity;
+//         }
+
+//         protected virtual void OnCreateCity(object sender, CreateCityEventArgs createCityArgs) {
+//             EntityManager entityManager = World.DefaultGameObjectInjectionWorld.EntityManager;
+
+//             Entity entity = SettlerEntity.GetSelected(entityManager);
+
+//             // SettlerEntity.AddCommandCreateCity(entityManager, entity, createCityArgs);
+//         }
+//     }
+// }
