+++ Assets/Scripts/UI/UIAssembly.asmdef
@@ -1,0 +1,18 @@
+{
+    "name": "UIAssembly",
+    "references": [
+        "GUID:734d92eba21c94caba915361bd5ac177",
+        "GUID:d8b63aba1907145bea998dd612889d6b",
+        "GUID:1f4f57f1c6b947e42a2fdacbf6936273",
+        "GUID:f5c81823401f2164484e02bfc181f689"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
